+++ qsp.h
@@ -8,7 +8,7 @@ uint8_t get10bitHighShift(uint8_t channel);
 uint8_t get10bitLowShift(uint8_t channel);
 void qspComputeCrc(QspConfiguration_t *qsp, uint8_t dataByte);
 void encodeRxHealthPayload(QspConfiguration_t *qsp, RxDeviceState_t *rxDeviceState, volatile RadioState_t *radioState);
-void encodeRcDataPayload(QspConfiguration_t *qsp, int channels[], uint8_t noOfChannels);
+void encodeRcDataPayload(QspConfiguration_t *qsp, int16_t channels[], uint8_t noOfChannels);
 void qspDecodeIncomingFrame(
     QspConfiguration_t *qsp, 
     uint8_t incomingByte,
